Extract model fetching logic to ModelManager class
